Beginning of the pseudo code syntax reader.
Converts a script line by line into a node of a tree based on the amount of tabs at the start of the line. "**BEGIN SCRIPT**" is the head of the tree. Hotkeys such as "CTRL + SHIFT + F" are children of the head, and scripts are children of the hotkeys. Lines inside while loops are children of while loops. Lines inside if statements are children of if statements, etc. Input validation (syntax) is only implemented for hotkeys atm. Commands such as highlight(), paste(), etc will come later.