Many things done (read desc)
StringSolver: similar to equation solver, takes in a string equation and returns a completed/combined string.
VariableDictionary: kind of unnecessary, but a reformat of python's dictionaries so that I can have an easier time using it
LinkedList: 2-way traversal linked list
CommandFormatter: takes in lines of pseudo code and translates them into values that are usable

only thing to do now for this sprint is hook up the original pesudo reader to the command formatter. if/while/for statements are not done, those are for another sprint.